refactor: 주문 페이지 컴포넌트 추상화 레벨 변경
